Change makefile

+++ ft_strncmp.c
@@ -6,14 +6,14 @@
 /*   By: ale-cont <ale-cont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 18:01:54 by ale-cont          #+#    #+#             */
-/*   Updated: 2022/11/09 20:18:36 by ale-cont         ###   ########.fr       */
+/*   Updated: 2022/11/10 13:36:21 by ale-cont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
